+++ plug-in-pool-mobile/app/src/main/java/com/example/plug_in_pool/ActiviteConfigurationMatch.java
@@ -1,10 +1,6 @@
 package com.example.plug_in_pool;
 
-import android.bluetooth.BluetoothAdapter;
-import android.bluetooth.BluetoothDevice;
-import android.bluetooth.BluetoothSocket;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -31,9 +27,6 @@ public class ActiviteConfigurationMatch extends AppCompatActivity
      */
     private static final String TAG =
       "_ActiviteConfigurationMatch"; //!< TAG pour les logs (cf. Logcat)
-    private static final int  DEMANDE_PERMISSIONS_BLUETOOTH = 1;
-    private static final UUID UUID_BLUETOOTH =
-      UUID.fromString("00001101-0000-1000-8000-00805F9B34FB");
 
     /**
      * Éléments de l'interface
@@ -41,15 +34,6 @@ public class ActiviteConfigurationMatch extends AppCompatActivity
     private Button               boutonLancerMatch;
     private AutoCompleteTextView choixNomJoueur1;
     private AutoCompleteTextView choixNomJoueur2;
-    private AutoCompleteTextView choixBluetoothEcran;
-    private AutoCompleteTextView choixBluetoothTable;
-
-    /**
-     * Bluetooth
-     */
-    private BluetoothSocket  socketBluetooth = null;
-    private BluetoothDevice  peripheriqueBluetooth;
-    private BluetoothAdapter adaptateurBluetooth;
 
     private BaseDeDonnees baseDonnees; //!< Classe d'accès avec la base de données
     Vector<Joueur>        joueurs = new Vector<Joueur>();
@@ -68,9 +52,6 @@ public class ActiviteConfigurationMatch extends AppCompatActivity
         Log.d(TAG, "onCreate()");
 
         initialiserVue();
-
-        demanderPermissionsBluetooth();
-        configurerBluetooth();
     }
 
     @Override
@@ -79,155 +60,20 @@ public class ActiviteConfigurationMatch extends AppCompatActivity
         super.onResume();
         Log.d(TAG, "onResume()");
 
-        if(ActivityCompat.checkSelfPermission(this,
-                                              android.Manifest.permission.BLUETOOTH_CONNECT) ==
-           PackageManager.PERMISSION_GRANTED)
-        {
-            listerPeripheriquesBluetooth();
-        }
-
         joueurs.clear();
     }
 
     private void initialiserVue()
     {
         Log.d(TAG, "initialiserVue()");
-        boutonLancerMatch   = findViewById(R.id.boutonLancerMatch);
-        choixNomJoueur1     = findViewById(R.id.choixNomJoueur1);
-        choixNomJoueur2     = findViewById(R.id.choixNomJoueur2);
-        choixBluetoothEcran = findViewById(R.id.choixBluetoothEcran);
-        choixBluetoothTable = findViewById(R.id.choixBluetoothTable);
+        boutonLancerMatch = findViewById(R.id.boutonLancerMatch);
+        choixNomJoueur1   = findViewById(R.id.choixNomJoueur1);
+        choixNomJoueur2   = findViewById(R.id.choixNomJoueur2);
 
         baseDonnees = BaseDeDonnees.getInstance(this);
 
         initialiserListeJoueurs();
         jouerMatch();
-    }
-
-    private void demanderPermissionsBluetooth()
-    {
-        if(android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.S)
-        {
-            if(ActivityCompat.checkSelfPermission(this,
-                                                  android.Manifest.permission.BLUETOOTH_CONNECT) !=
-                 PackageManager.PERMISSION_GRANTED ||
-               ActivityCompat.checkSelfPermission(this,
-                                                  android.Manifest.permission.BLUETOOTH_SCAN) !=
-                 PackageManager.PERMISSION_GRANTED)
-            {
-                ActivityCompat.requestPermissions(
-                  this,
-                  new String[] { android.Manifest.permission.BLUETOOTH_CONNECT,
-                                 android.Manifest.permission.BLUETOOTH_SCAN },
-                  DEMANDE_PERMISSIONS_BLUETOOTH);
-            }
-        }
-    }
-
-    private void configurerBluetooth()
-    {
-        adaptateurBluetooth = BluetoothAdapter.getDefaultAdapter();
-
-        if(adaptateurBluetooth == null)
-        {
-            Log.e(TAG, "Bluetooth non supporté sur cet appareil");
-            return;
-        }
-        if(!adaptateurBluetooth.isEnabled())
-        {
-            Intent activerBluetooth = new Intent(BluetoothAdapter.ACTION_REQUEST_ENABLE);
-            startActivityForResult(activerBluetooth, 1);
-        }
-    }
-
-    private void listerPeripheriquesBluetooth()
-    {
-        if(adaptateurBluetooth == null)
-        {
-            Log.e(TAG, "Bluetooth non disponible !");
-            return;
-        }
-
-        if(ActivityCompat.checkSelfPermission(this,
-                                              android.Manifest.permission.BLUETOOTH_CONNECT) !=
-           PackageManager.PERMISSION_GRANTED)
-        {
-            Log.e(TAG, "Permissions Bluetooth refusées !");
-            return;
-        }
-
-        Set<BluetoothDevice> appareilsAppaires = adaptateurBluetooth.getBondedDevices();
-        List<String>         nomsAppareils     = new ArrayList<>();
-
-        for(BluetoothDevice appareil: appareilsAppaires)
-        {
-            nomsAppareils.add(appareil.getName() + " (" + appareil.getAddress() + ")");
-        }
-
-        ArrayAdapter<String> adaptateur =
-          new ArrayAdapter<>(this, android.R.layout.simple_dropdown_item_1line, nomsAppareils);
-        choixBluetoothEcran.setAdapter(adaptateur);
-        choixBluetoothTable.setAdapter(adaptateur);
-
-        choixBluetoothEcran.setOnItemClickListener((parent, view, position, id) -> {
-            String selection  = (String)parent.getItemAtPosition(position);
-            String adresseMac = extraireAdresseMac(selection);
-            connecterBluetooth(adresseMac);
-        });
-
-        choixBluetoothTable.setOnItemClickListener((parent, view, position, id) -> {
-            String selection  = (String)parent.getItemAtPosition(position);
-            String adresseMac = extraireAdresseMac(selection);
-            connecterBluetooth(adresseMac);
-        });
-    }
-
-    private void connecterBluetooth(String adresseMac)
-    {
-        if(ActivityCompat.checkSelfPermission(this,
-                                              android.Manifest.permission.BLUETOOTH_CONNECT) !=
-           PackageManager.PERMISSION_GRANTED)
-        {
-            Log.e(TAG, "Permissions Bluetooth refusées !");
-            return;
-        }
-
-        peripheriqueBluetooth = adaptateurBluetooth.getRemoteDevice(adresseMac);
-        Log.d(TAG, "Tentative de connexion à l'adresse MAC : " + adresseMac);
-
-        try
-        {
-            Log.d(TAG, "Création du socket Bluetooth...");
-            socketBluetooth =
-              peripheriqueBluetooth.createInsecureRfcommSocketToServiceRecord(UUID_BLUETOOTH);
-
-            Log.d(TAG, "Socket créée, démarrage de la connexion...");
-            socketBluetooth.connect();
-            Log.d(TAG, "Connexion Bluetooth réussie à " + adresseMac);
-        }
-        catch(IOException e)
-        {
-            Log.e(TAG, "Échec de connexion (méthode normale) : " + e.getMessage());
-
-            try
-            {
-                Log.d("Bluetooth", "Tentative de connexion alternative...");
-                socketBluetooth = (BluetoothSocket)peripheriqueBluetooth.getClass()
-                                    .getMethod("createRfcommSocket", new Class[] { int.class })
-                                    .invoke(peripheriqueBluetooth, 1);
-                socketBluetooth.connect();
-                Log.d(TAG, "Connexion alternative réussie !");
-            }
-            catch(Exception ex)
-            {
-                Log.e(TAG, "Échec de la connexion alternative : " + ex.getMessage());
-            }
-        }
-    }
-
-    private String extraireAdresseMac(String texte)
-    {
-        return texte.substring(texte.lastIndexOf('(') + 1, texte.length() - 1);
     }
 
     private void initialiserListeJoueurs()
@@ -240,9 +86,14 @@ public class ActiviteConfigurationMatch extends AppCompatActivity
         choixNomJoueur2.setAdapter(listeJoueurs);
     }
 
-    private boolean ajouterJoueur1()
+    private boolean ajouterJoueur(int numeroJoueur)
     {
-        String nomJoueur       = choixNomJoueur1.getText().toString().trim();
+        String nomJoueur = "";
+        if(numeroJoueur == 1)
+            nomJoueur = choixNomJoueur1.getText().toString().trim();
+        else if(numeroJoueur == 2)
+            nomJoueur = choixNomJoueur2.getText().toString().trim();
+
         String[] donneesJoueur = nomJoueur.split(" ");
         if(donneesJoueur.length == 2)
         {
@@ -250,25 +101,9 @@ public class ActiviteConfigurationMatch extends AppCompatActivity
             String nom    = donneesJoueur[1];
             Joueur joueur = new Joueur(nom, prenom);
             joueurs.add(joueur);
-            Log.d(TAG, "ajouterJoueur1() " + joueur.getPrenom() + " " + joueur.getNom());
-            baseDonnees.ajouterJoueur(nom, prenom);
-            return true;
-        }
-
-        return false;
-    }
-
-    private boolean ajouterJoueur2()
-    {
-        String nomJoueur       = choixNomJoueur2.getText().toString().trim();
-        String[] donneesJoueur = nomJoueur.split(" ");
-        if(donneesJoueur.length == 2)
-        {
-            String prenom = donneesJoueur[0];
-            String nom    = donneesJoueur[1];
-            Joueur joueur = new Joueur(nom, prenom);
-            joueurs.add(joueur);
-            Log.d(TAG, "ajouterJoueur2() " + joueur.getPrenom() + " " + joueur.getNom());
+            Log.d(TAG,
+                  "ajouterJoueur(" + numeroJoueur + ") " + joueur.getPrenom() + " " +
+                    joueur.getNom());
             baseDonnees.ajouterJoueur(nom, prenom);
             return true;
         }
@@ -278,39 +113,41 @@ public class ActiviteConfigurationMatch extends AppCompatActivity
 
     private int getValeurSaisie()
     {
-        EditText nbParties = findViewById(R.id.choixNbParties);
-        String choixUtilisateur = nbParties.getText().toString().trim();
+        EditText nbParties        = findViewById(R.id.choixNbParties);
+        String   choixUtilisateur = nbParties.getText().toString().trim();
 
-        if (choixUtilisateur.isEmpty())
+        if(choixUtilisateur.isEmpty())
         {
             return 1;
         }
 
-        try {
+        try
+        {
             return Integer.parseInt(choixUtilisateur);
-        } catch (NumberFormatException e)
+        }
+        catch(NumberFormatException e)
         {
             Log.d("getValeurSaisie", "Entrée non valide : " + choixUtilisateur);
             return 1;
         }
     }
-    private void ajoutsBaseDeDonnees()
+
+    private boolean enregistrerDonnees()
     {
-        if (!ajouterLesDeuxJoueurs())
+        if(!ajouterLesDeuxJoueurs())
         {
-            Log.d(TAG, "Erreur : Les deux joueurs doivent être correctement saisis (prénom nom).");
-            return;
+            return false;
         }
 
-        creerEtLoggerLeMatch();
+        return creerLeMatch();
     }
 
     private boolean ajouterLesDeuxJoueurs()
     {
-        return ajouterJoueur1() && ajouterJoueur2();
+        return ajouterJoueur(1) && ajouterJoueur(2);
     }
 
-    private void creerEtLoggerLeMatch()
+    private boolean creerLeMatch()
     {
         int nbParties = getValeurSaisie();
 
@@ -320,14 +157,16 @@ public class ActiviteConfigurationMatch extends AppCompatActivity
         int id1 = baseDonnees.getIdJoueur(joueur1.getNom(), joueur1.getPrenom());
         int id2 = baseDonnees.getIdJoueur(joueur2.getNom(), joueur2.getPrenom());
 
-        if (id1 != -1 && id2 != -1)
+        Log.d(TAG, "id1 : " + id1 + ", id2 : " + id2 + ", nbParties : " + nbParties);
+
+        if(id1 != -1 && id2 != -1)
         {
             baseDonnees.creerMatch(id1, id2, nbParties);
-            Log.d(TAG, "id1 : " + id1 + ", id2 : " + id2 + ", Nombre de parties : " + nbParties);
+            return true;
         }
         else
         {
-            Log.d(TAG, "Un des joueurs est introuvable dans la base !");
+            return false;
         }
     }
 
@@ -342,17 +181,20 @@ public class ActiviteConfigurationMatch extends AppCompatActivity
                     Log.d(TAG,
                           "jouerMatch() " + choixNomJoueur1.getText().toString() + " vs " +
                             choixNomJoueur2.getText().toString());
-                    Log.d(TAG, "jouerMatch() écran : " + choixBluetoothEcran.getText().toString());
-                    Log.d(TAG, "jouerMatch() table : " + choixBluetoothTable.getText().toString());
-                    ajoutsBaseDeDonnees();
-                    int nbParties = getValeurSaisie();
-                    Log.d(TAG, "jouerMatch() nbJoueurs : " + joueurs.size());
-                    Intent changerVue =
-                      new Intent(ActiviteConfigurationMatch.this, ActiviteGestionPartie.class);
-                    changerVue.putExtra("joueur1", joueurs.elementAt(0));
-                    changerVue.putExtra("joueur2", joueurs.elementAt(1));
-                    changerVue.putExtra("nbParties", nbParties);
-                    startActivity(changerVue);
+
+                    if(enregistrerDonnees())
+                    {
+                        Log.d(TAG, "jouerMatch() nbJoueurs : " + joueurs.size());
+                        Intent changerVue =
+                          new Intent(ActiviteConfigurationMatch.this, ActiviteGestionPartie.class);
+                        changerVue.putExtra("joueur1", joueurs.elementAt(0));
+                        changerVue.putExtra("joueur2", joueurs.elementAt(1));
+                        changerVue.putExtra("nbParties", getValeurSaisie());
+                        startActivity(changerVue);
+                    }
+                    else
+                    {
+                    }
                 }
                 else
                 {
