+++ plug-in-pool-mobile/app/src/main/java/com/example/plug_in_pool/CommunicationBluetooth.java
@@ -1,0 +1,4 @@
+package com.example.plug_in_pool;
+
+public class CommunicationBluetooth {
+}
