+++ plug-in-pool-mobile/app/src/main/res/layout/activity_configuration_match.xml
@@ -17,7 +17,7 @@
         app:layout_constraintEnd_toEndOf="parent">
 
         <TextView
-            android:id="@+id/textView2"
+            android:id="@+id/textViewNomApplication"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:fontFamily="sans-serif-medium"
@@ -33,7 +33,7 @@
             android:paddingEnd="16dp"/>
 
         <ImageView
-            android:id="@+id/imageView"
+            android:id="@+id/imageViewLogoApplication"
             android:layout_width="270dp"
             android:layout_height="149dp"
             android:contentDescription="@string/logoCampus"
@@ -125,89 +125,11 @@
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/liaisonBluetooth"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf="@id/choixJoueur"
-        android:layout_marginTop="30dp">
-
-        <TextView
-            android:id="@+id/nomBluetoothEcran"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:contentDescription="@string/texteChoixBluetoothEcran"
-            android:fontFamily="sans-serif-medium"
-            android:gravity="center_horizontal|center_vertical"
-            android:paddingStart="16dp"
-            android:paddingEnd="16dp"
-            android:text="@string/texteChoixBluetoothEcran"
-            android:textColor="@color/vert"
-            android:textSize="20sp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@id/nomBluetoothTable"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/nomBluetoothTable"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:contentDescription="@string/texteChoixBluetoothTable"
-            android:fontFamily="sans-serif-medium"
-            android:gravity="center_horizontal|center_vertical"
-            android:paddingStart="16dp"
-            android:paddingEnd="16dp"
-            android:text="@string/texteChoixBluetoothTable"
-            android:textColor="@color/vert"
-            android:textSize="20sp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/nomBluetoothEcran"
-            app:layout_constraintTop_toTopOf="parent" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/choixBluetooth"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="10dp"
-        android:layout_marginEnd="10dp"
-        android:layout_marginTop="10dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/liaisonBluetooth">
-
-        <AutoCompleteTextView
-            android:id="@+id/choixBluetoothEcran"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:hint="@string/choisirLiaisonBluetooth"
-            android:paddingTop="12dp"
-            android:paddingBottom="12dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@id/choixBluetoothTable"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <AutoCompleteTextView
-            android:id="@+id/choixBluetoothTable"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:hint="@string/choisirLiaisonBluetooth"
-            android:paddingTop="12dp"
-            android:paddingBottom="12dp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/choixBluetoothEcran" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/nbParties"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="60dp"
-        app:layout_constraintTop_toBottomOf="@id/choixBluetooth"
+        app:layout_constraintTop_toBottomOf="@id/choixJoueur"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent">
