+++ plug-in-pool-mobile/app/src/main/res/values/strings.xml
@@ -9,9 +9,10 @@
     <string name="texteSelectionJoueur2">Joueur n°2</string>
     <string name="choisirUnJoueurEnregistrer1">Saisir le nom du joueur et/ou choisir dans la liste</string>
     <string name="choisirUnJoueurEnregistrer2">Saisir le nom du joueur et/ou choisir dans la liste</string>
+    <string name="nbParties">Choisir le nombre de parties</string>
     <string name="texteChoixBluetoothEcran">Choisir un écran</string>
     <string name="texteChoixBluetoothTable">Choisir une table</string>
     <string name="choisirLiaisonBluetooth">Saisir le flux Bluetooth et choisir dans la liste</string>
     <string name="boutonLancerMatch">Lancer un match</string>
-    <string name="nbParties">Choisir le nombre de parties</string>
+    <string name="boutonLancerPartie">Lancer une partie</string>
 </resources>
